# specs/001-agent-tutor/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: AgentTutor Backend API
  description: |
    Zero-Backend-LLM deterministic API for the AgentTutor Course Companion FTE.
    Phase 1: No LLM calls. All intelligence handled by ChatGPT App frontend.
  version: "1.0.0"
  contact:
    name: AgentTutor Team
servers:
  - url: https://agenttutor-backend.fly.dev
    description: Production (Fly.io)
  - url: http://localhost:8000
    description: Local development
components:
  parameters:
    UserId:
      name: user_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Student UUID identifier
  schemas:
...
    UserCreate:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          example: student@example.com
        name:
          type: string
          example: Ahmed Khan
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        tier:
          type: string
          enum: [free, premium]
        streak_days:
          type: integer
          minimum: 0
        last_active:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Chapters ───────────────────────────────────────────────────────────
...
    SearchResult:
      type: object
      properties:
        chapter_id:
          type: string
        chapter_title:
          type: string
        chapter_number:
          type: integer
        excerpt:
          type: string
          description: Up to 300-character excerpt around the matching keyword

    # ── Quizzes ────────────────────────────────────────────────────────────
    Question:
      type: object
      properties:
        question_number:
          type: integer
        question:
          type: string
        options:
          type: object
          properties:
            A:
              type: string
            B:
              type: string
            C:
              type: string
            D:
              type: string
    QuizQuestions:
      type: object
      properties:
        chapter_id:
          type: string
        chapter_title:
          type: string
        total_questions:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          description: Map of question_number (as string) to selected option
          additionalProperties:
            type: string
            enum: [A, B, C, D]
          example:
            "1": "B"
            "2": "A"
            "3": "C"
            "4": "D"
            "5": "A"
    QuizResult:
      type: object
      properties:
        chapter_id:
          type: string
        score:
          type: integer
          description: Number of correct answers
        total:
          type: integer
          description: Total questions
        percentage:
          type: integer
          description: Score as percentage (0–100)
        passed:
          type: boolean
          description: True if score >= 4 (80%)
        wrong_question_numbers:
          type: array
          items:
            type: integer
          description: List of question numbers answered incorrectly

    # ── Progress ───────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: premium_required
        message:
          type: string
          example: "This chapter requires a premium subscription."
        detail:
          type: string
          nullable: true
paths:
...
  /users:
    post:
      operationId: upsert_user
      summary: Create or retrieve user by email
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        "200":
          description: Existing user returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "201":
          description: New user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    get:
      operationId: get_user
      summary: Get user profile by ID
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Chapters ───────────────────────────────────────────────────────────
...
  /search:
    get:
      operationId: search_content
      summary: Keyword search across accessible chapter content
      tags: [Search]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Keyword to search
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        "200":
          description: Matching chapter excerpts (empty array if no matches)
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
                  total:
                    type: integer

  # ── Quizzes ────────────────────────────────────────────────────────────
  /quizzes/{chapter_id}/questions:
    get:
      operationId: get_quiz_questions
      summary: Get quiz questions for a chapter (options only, no answers)
      tags: [Quizzes]
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Quiz questions without correct answer revealed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizQuestions'
        "403":
          description: Premium quiz, free-tier user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No quiz found for chapter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quizzes/{chapter_id}/submit:
    post:
      operationId: submit_quiz
      summary: Submit quiz answers for rule-based grading
      tags: [Quizzes]
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizSubmission'
      responses:
        "200":
          description: Graded result (no correct answers revealed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResult'
        "403":
          description: Premium quiz, free-tier user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Progress ───────────────────────────────────────────────────────────
...